backend/generate-image/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
boto3>=1.28.0
